miscellaneous fixes, add some basic tests
