+++ docs/index.html
@@ -14,9 +14,9 @@
 
         <!-- config for our rust wasm binary. go to https://trunkrs.dev/assets/#rust for more customization -->
         
-<script type="module" nonce="CfGfH9NLLZnx0sk7ABvxFw==">
-import init, * as bindings from '/raphael-xiv-8c5ee62c7b03f61d.js';
-const wasm = await init('/raphael-xiv-8c5ee62c7b03f61d_bg.wasm');
+<script type="module" nonce="ruz2f3BP56583PIcMJFuqg==">
+import init, * as bindings from '/raphael-xiv-bddc71701287ad17.js';
+const wasm = await init('/raphael-xiv-bddc71701287ad17_bg.wasm');
 
 
 window.wasmBindings = bindings;
@@ -81,8 +81,8 @@ dispatchEvent(new CustomEvent("TrunkApplicationStarted", {detail: {wasm}}));
             }
         </style>
     
-<link rel="modulepreload" href="/raphael-xiv-8c5ee62c7b03f61d.js" crossorigin=anonymous integrity="sha384-iEXH0D42vc5fA6FIagk5KGCrIWlsMOa+cU4C+w0P0CRpzHxRuCmEOUwqDzKKd80J">
-<link rel="preload" href="/raphael-xiv-8c5ee62c7b03f61d_bg.wasm" crossorigin=anonymous integrity="sha384-BZu3QAUuUo8ajYyOVn8KKIQ1wvZBHu1rjr8vtCWOOZCV8yEfPG/dbvCiYvEKiLBi" as="fetch" type="application/wasm"></head>
+<link rel="modulepreload" href="/raphael-xiv-bddc71701287ad17.js" crossorigin=anonymous integrity="sha384-HfLiKJJFPfzRZPg4nr7HOzY4+qtcSrd8LW0cCkEs6Xv4vbP4+mnUIIJPINEdWD97">
+<link rel="preload" href="/raphael-xiv-bddc71701287ad17_bg.wasm" crossorigin=anonymous integrity="sha384-ENfpmjcBAGsccJfXQ9lnrxWqAErefvyt1f/n84MXJNF0Hshas1Wg6wtrEZ/3vAkh" as="fetch" type="application/wasm"></head>
 
     <body>
         <!-- The WASM code will resize the canvas dynamically -->
